// frontend/packages/web/src/enums/DashboardWidgetPeriod.ts
export enum DashboardWidgetPeriod {
  BiWeek = 'BI_WEEK',
  Custom = 'CUSTOM',
  Month = 'MONTH',
  Quarter = 'QUARTER',
  QuarterPerDay = 'QUARTER_PER_DAY',
  QuarterPerWeek = 'QUARTER_PER_WEEK',
  Semester = 'SEMESTER',
  Week = 'WEEK',
  Year = 'YEAR',
  YearPerMonth = 'YEAR_PER_MONTH',
  YearPerWeek = 'YEAR_PER_WEEK',
}